# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/resume-upload": {
            "get": {
                "description": "Check uploaded chunks and return status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "Resume Upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File name",
                        "name": "file_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResumeUploadResponse"
                        }
                    }
                }
            }
        },
        "/upload": {
            "get": {
                "description": "Serve the upload HTML page",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "Upload Page",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/upload-chunk": {
            "post": {
                "description": "Handle file chunk upload",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "Upload Chunk",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File chunk",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Chunk index",
                        "name": "chunk_index",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Total chunks",
                        "name": "total_chunks",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File name",
                        "name": "file_name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.UploadChunkResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ResumeUploadResponse": {
            "type": "object",
            "properties": {
                "chunk_index": {
                    "type": "integer"
                },
                "file_name": {
                    "type": "string"
                },
                "total_chunks": {
                    "type": "integer"
                }
            }
        },
        "handlers.UploadChunkResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        }
    }
}